refactor: re-organize tier and site modules to prep for middleware
 - move `lms/djangoapps/appsembler_tiers/` into `openedx/core/djangoapps/appsembler/tahoe_tiers/`
 - keeping `cms/djangoapps/appsembler_tiers/` in its place due to test complications
 - move `get_active_site_uuids_from_site_config_service` into `site_config_client_helpers`
 - move `get_active_tiers_uuids_from_amc_postgres` into
   `legacy_amc_helpers`
 - move TiersDbRouter into `appsembler/tahoe_tiers`

This commit is expected to do _nothing_ except for refactoring, another commit will include the needed fixes.
