--- lms/djangoapps/appsembler_tiers/__init__.py
@@ -1,5 +1,0 @@
-"""
-Views and urls related to the `tiers` app.
-
-This app gets redirects to the Studio `appsembler_tiers` app.
-"""
