+++ openedx/core/djangoapps/appsembler/settings/settings/production_common.py
@@ -67,7 +67,7 @@ def plugin_settings(settings):
 
         settings.TIERS_DATABASE_URL = settings.AUTH_TOKENS.get('TIERS_DATABASE_URL')
         settings.DATABASES['tiers'] = dj_database_url.parse(settings.TIERS_DATABASE_URL, ssl_require=True)
-        settings.DATABASE_ROUTERS.insert(0, 'openedx.core.djangoapps.appsembler.sites.routers.TiersDbRouter')
+        settings.DATABASE_ROUTERS.insert(0, 'openedx.core.djangoapps.appsembler.tahoe_tiers.db_routers.TiersDbRouter')
 
         settings.MIDDLEWARE += [
             'tiers.middleware.TierMiddleware',
