+++ openedx/core/djangoapps/appsembler/sites/utils.py
@@ -1,6 +1,8 @@
-from uuid import UUID
+"""
+A big module for Tahoe Sites multi-tenancy helpers.
 
-import logging
+A lot of this module should be migrated into more specific modules such as `tahoe-sites`.
+"""
 
 from datetime import timedelta
 
@@ -12,7 +14,7 @@ import cssutils
 import os
 import sass
 
-from django.db.models import Q, F
+from django.db.models import Q
 from django.utils import timezone
 from django.core.files.storage import get_storage_class
 from django.conf import settings
@@ -35,10 +37,8 @@ from openedx.core.lib.log_utils import audit_log
 from openedx.core.djangoapps.theming.helpers import get_current_request, get_current_site
 from openedx.core.djangoapps.theming.models import SiteTheme
 
-from site_config_client.exceptions import SiteConfigurationError
-
-
-log = logging.getLogger(__name__)
+from ..tahoe_tiers.legacy_amc_helpers import get_active_tiers_uuids_from_amc_postgres
+from .site_config_client_helpers import get_active_site_uuids_from_site_config_service
 
 
 @beeline.traced(name="get_lms_link_from_course_key")
@@ -56,48 +56,6 @@ def get_lms_link_from_course_key(base_lms_url, course_key):
     if course_site:
         return course_site.domain
     return base_lms_url
-
-
-def get_active_tiers_uuids_from_amc_postgres():
-    """
-    Get active Tier organization UUIDs from the Tiers (AMC Postgres) database.
-
-    Note: This mostly a hack that's needed for improving the performance of
-          batch operations by excluding dead sites.
-
-    Return a list of UUID objects.
-
-    TODO: This helper should live in a future Tahoe Sites package.
-    """
-    from tiers.models import Tier
-    # This queries the AMC Postgres database
-    active_tiers_uuids = Tier.objects.filter(
-        Q(tier_enforcement_exempt=True) |
-        Q(tier_expires_at__gte=timezone.now())
-    ).annotate(
-        organization_edx_uuid=F('organization__edx_uuid')
-    ).values_list('organization_edx_uuid', flat=True)
-    return list(active_tiers_uuids)
-
-
-def get_active_site_uuids_from_site_config_service():
-    """
-    Get active Tier organization UUIDs via the client of the Site Configuration service.
-
-    Return a list of UUID objects.
-    TODO: Move this helper into another package.
-    """
-    client = getattr(settings, 'SITE_CONFIG_CLIENT', None)
-    if client:
-        try:
-            active_sites_response = client.list_active_sites()
-            active_sites = active_sites_response['results']
-            site_uuids = [UUID(site['uuid']) for site in active_sites]
-            return site_uuids
-        except SiteConfigurationError:
-            log.exception('An error occurred while fetching site config active sites, returning an empty list.')
-
-    return []
 
 
 def get_active_organizations_uuids():
