+++ openedx/core/djangoapps/appsembler/tahoe_tiers/__init__.py
@@ -1,0 +1,5 @@
+"""
+Tahoe complements of to the `django-tiers` package.
+
+TODO: This app is temporary until AMC is shut down, then it probably should be moved into a new `tahoe-tiers` package.
+"""
