+++ tox.ini
@@ -114,7 +114,6 @@ commands =
     pytest {env:PYTEST_ARGS} \
         common/djangoapps/util/tests/test_db.py::MigrationTests  \
         lms/tests.py \
-        lms/djangoapps/appsembler_tiers \
         lms/djangoapps/certificates/tests/test_webview_appsembler_changes.py  \
         lms/djangoapps/course_api/ \
         lms/djangoapps/course_blocks/transformers/tests/test_load_override_data.py \
